# docs/swagger.yaml
definitions:
...
  core_dtos.EmailDto:
    properties:
      email:
        type: string
      status:
        type: string
    type: object
  core_dtos.GetUserResponseDto:
    properties:
      calendar_settings:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        items:
          $ref: '#/definitions/core_dtos.EmailDto'
        type: array
      first_name:
        type: string
      google_id:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_verified:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      locale:
        type: string
      notification_settings:
        type: string
      profile_picture:
        type: string
      role:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
    type: object
...
  core_dtos.UpdateProfileRequestDto:
    properties:
      calendar_settings:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      notification_settings:
        type: string
      profile_picture:
        type: string
    type: object
...
  fiber.Map:
    additionalProperties: true
    type: object
...
info:
  contact: {}
  description: Timewise API
  title: timewise-api
  version: "1.0"
paths:
  /api/v1/account/user:
    get:
      consumes:
      - application/json
      description: Get user info
      parameters:
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core_dtos.GetUserResponseDto'
      security:
      - bearerToken: []
      summary: Get user info
      tags:
      - account
    patch:
      consumes:
      - application/json
      description: Update user info
      parameters:
      - description: Update user info request
        in: body
        name: updateUserInfoRequest
        required: true
        schema:
          $ref: '#/definitions/core_dtos.UpdateProfileRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core_dtos.GetUserResponseDto'
      security:
      - bearerToken: []
      summary: Update user info
      tags:
      - account
  /api/v1/account/user/deactivate:
    post:
      consumes:
      - application/json
      description: Deactivate account
      produces:
      - application/json
      responses:
        "200":
          description: Account deactivated
          schema:
            type: string
        "400":
          description: Invalid userId
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - bearerToken: []
      summary: Deactivate account
      tags:
      - account
  /api/v1/account/user/emails:
    get:
      consumes:
      - application/json
      description: Get linked user emails
      parameters:
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - bearerToken: []
      summary: Get linked user emails
      tags:
      - account
  /api/v1/account/user/emails/link/{token}:
    get:
      consumes:
      - application/json
      description: Action email link request
      parameters:
      - description: Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core_dtos.GetUserResponseDto'
      summary: Action email link request
      tags:
      - account
  /api/v1/account/user/emails/send:
    post:
      consumes:
      - application/json
      description: Send link an email request
      parameters:
      - description: Target email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Link email request sent
          schema:
            type: string
      security:
      - bearerToken: []
      summary: Send link an email request
      tags:
      - account
  /api/v1/account/user/emails/unlink:
    post:
      consumes:
      - application/json
      description: Unlink an email
      parameters:
      - description: Email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core_dtos.GetUserResponseDto'
      security:
      - bearerToken: []
      summary: Unlink an email
      tags:
      - account
...
securityDefinitions:
  bearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"